+++ src/test/ui/parser/macro/bad-macro-argument.stderr
